refactor(core): extract movementProfile helpers from Publisher (no behavior change)
Lift the yaml-mutex + profile-add/remove + sync-twin + signal-handler
infrastructure from core/Publisher.ts into a new internal module,
core/movementProfile.ts. Publisher.ts now imports the helpers instead
of owning them.

Why: M2.2 (#117) needs the same profile-management + SIGINT-safe
cleanup for `move deploy-object` / `move upgrade-object`. Duplicating
the bug-#36 / #37 / #43 hardening across two flows would invite drift.
A standalone helper module is the smallest move that lets both
`Publisher` (publish) and the upcoming `harness/codeObject` (deploy-
object) share the hardening without coupling them.

The extraction is purely mechanical:
  - 9 functions / interfaces / module-state vars moved verbatim
  - Publisher.ts goes from 491 LoC to ~325 LoC
  - Publisher's class body is untouched; only the imports change

Verified behavior-neutral: 200/200 unit tests pass unchanged,
including the concurrent-deploy + signal-handler suites in
__tests__/deployContract.test.ts that exercise withYamlLock and
the cleanupCallbacks set transitively.

Tracks #69. Step 1 of 2 for M2.2 (#117).
